+++ ext/QuantumCliffordHeckeExt/lifted_product.jl
@@ -165,9 +165,9 @@ function _build_parity_checks(::NonCommutativeLift, c::LPCode)
     return hx, hz
 end
 
-parity_checks_x(c::LPCode) = parity_checks_xz(c)[1]
+parity_matrix_x(c::LPCode) = parity_checks_xz(c)[1]
 
-parity_checks_z(c::LPCode) = parity_checks_xz(c)[2]
+parity_matrix_z(c::LPCode) = parity_checks_xz(c)[2]
 
 parity_checks(c::LPCode) = parity_checks(CSS(parity_checks_xz(c)...))
 
@@ -199,8 +199,10 @@ julia> B = 1 + x^7 + s + x^8 + s*x^7 + x;
 
 julia> c = two_block_group_algebra_codes(A,B);
 
-julia> code_n(c), code_k(c)
-(56, 28)
+julia> import HiGHS;
+
+julia> code_n(c), code_k(c), distance(c, DistanceMIPAlgorithm(solver=HiGHS))
+(56, 28, 2)
 ```
 
 ### Bivariate Bicycle codes
@@ -256,8 +258,10 @@ julia> B = x + z^5 + y^5 + y^2;
 
 julia> c = two_block_group_algebra_codes(A, B);
 
-julia> code_n(c), code_k(c)
-(48, 4)
+julia> import HiGHS
+
+julia> code_n(c), code_k(c), distance(c, DistanceMIPAlgorithm(solver=HiGHS))
+(48, 4, 6)
 ```
 
 ### Coprime Bivariate Bicycle code
@@ -284,11 +288,13 @@ julia> B = 𝜋^8 + 𝜋^14 + 𝜋^47;
 
 julia> c = two_block_group_algebra_codes(A, B);
 
-julia> code_n(c), code_k(c)
-(108, 12)
+julia> import HiGHS
+
+julia> code_n(c), code_k(c), distance(c, DistanceMIPAlgorithm(solver=HiGHS))
+(108, 12, 6)
 ```
 
-See also: [`LPCode`](@ref), [`generalized_bicycle_codes`](@ref), [`bicycle_codes`](@ref), [`haah_cubic_codes`](@ref).
+See also: [`QuantumClifford.ECC.LPCode`](@ref), [`generalized_bicycle_codes`](@ref), [`bicycle_codes`](@ref), [`haah_cubic_codes`](@ref).
 """
 function two_block_group_algebra_codes(a::GroupAlgebraElem, b::GroupAlgebraElem)
     LPCode([a;;], [b;;])