+++ test/test_ecc_multivariate_bicycle.jl
@@ -1,7 +1,9 @@
 @testitem "ECC Multivaraite Bicycle" begin
     using Hecke
+    using HiGHS
+    using JuMP
     using Hecke: group_algebra, GF, abelian_group, gens
-    using QuantumClifford.ECC: two_block_group_algebra_codes, code_k, code_n
+    using QuantumClifford.ECC: two_block_group_algebra_codes, code_k, code_n, distance, DistanceMIPAlgorithm
 
     # Multivariate Bicycle codes taken from Table 1 of [voss2024multivariatebicyclecodes](@cite)
     @testset "Weight-4 QLDPC Codes" begin
@@ -14,6 +16,7 @@
         B = x + x^6
         c = two_block_group_algebra_codes(A, B)
         @test code_n(c) == 112 && code_k(c) == 8
+        @test distance(c, DistanceMIPAlgorithm(solver=HiGHS)) == 5
 
         # [[64, 2, 8]]
         l=8; m=4
@@ -24,6 +27,7 @@
         B = x^3 + y
         c = two_block_group_algebra_codes(A, B)
         @test code_n(c) == 64 && code_k(c) == 2
+        @test distance(c, DistanceMIPAlgorithm(solver=HiGHS)) == 8
 
         # [[72, 2, 8]]
         l=4; m=9
@@ -34,6 +38,7 @@
         B = x^2 + y^2
         c = two_block_group_algebra_codes(A, B)
         @test code_n(c) == 72 && code_k(c) == 2
+        @test distance(c, DistanceMIPAlgorithm(solver=HiGHS)) == 8
 
         # [[96, 2, 8]]
         l=6; m=8
@@ -44,6 +49,7 @@
         B = z + z^4
         c = two_block_group_algebra_codes(A, B)
         @test code_n(c) == 96 && code_k(c) == 2
+        @test distance(c, DistanceMIPAlgorithm(solver=HiGHS)) == 8
 
         # [[112, 2, 10]]
         l=7; m=8
@@ -54,6 +60,7 @@
         B = z^2 + y^5
         c = two_block_group_algebra_codes(A, B)
         @test code_n(c) == 112 && code_k(c) == 2
+        @test distance(c, DistanceMIPAlgorithm(solver=HiGHS)) == 10
 
         # [[144, 2, 12]]
         l=8; m=9
@@ -64,6 +71,7 @@
         B = x + y^5
         c = two_block_group_algebra_codes(A, B)
         @test code_n(c) == 144 && code_k(c) == 2
+        @test distance(c, DistanceMIPAlgorithm(solver=HiGHS)) == 12
     end
 
     @testset "Weight-5 QLDPC Codes" begin
@@ -76,6 +84,7 @@
         B = x + y^2 + z^2 
         c = two_block_group_algebra_codes(A, B)
         @test code_n(c) == 30 && code_k(c) == 4
+        @test distance(c, DistanceMIPAlgorithm(solver=HiGHS)) == 5
 
         # [[72, 4, 8]]
         l=4; m=9
@@ -86,6 +95,7 @@
         B = x^2 + y + y^2 
         c = two_block_group_algebra_codes(A, B)
         @test code_n(c) == 72 && code_k(c) == 4
+        @test distance(c, DistanceMIPAlgorithm(solver=HiGHS)) == 8
 
         # [96, 4, 8]]
         l=8; m=6
@@ -96,6 +106,7 @@
         B = z^5 + x^5 + y 
         c = two_block_group_algebra_codes(A, B)
         @test code_n(c) == 96 && code_k(c) == 4
+        @test distance(c, DistanceMIPAlgorithm(solver=HiGHS)) == 8
     end
 
     @testset "Weight-6 QLDPC codes" begin
@@ -108,6 +119,7 @@
         B = x^4 + x + z^4 + y
         c = two_block_group_algebra_codes(A, B)
         @test code_n(c) == 30 && code_k(c) == 6
+        @test distance(c, DistanceMIPAlgorithm(solver=HiGHS)) == 4
 
         # [[48, 6, 6]]
         l=4; m=6
@@ -118,6 +130,7 @@
         B = x^3 + z^3 + y^2 + y
         c = two_block_group_algebra_codes(A, B)
         @test code_n(c) == 48 && code_k(c) == 6
+        @test distance(c, DistanceMIPAlgorithm(solver=HiGHS)) == 6
 
         # [[40, 4, 6]]
         l=4; m=5
@@ -128,6 +141,7 @@
         B = y^4 + y^2 + x^3 + x
         c = two_block_group_algebra_codes(A, B)
         @test code_n(c) == 40 && code_k(c) == 4
+        @test distance(c, DistanceMIPAlgorithm(solver=HiGHS)) == 6
 
         # [[48, 4, 6]]
         l=4; m=6
@@ -138,6 +152,7 @@
         B = x + z^5 + y^5 + y^2
         c = two_block_group_algebra_codes(A, B)
         @test code_n(c) == 48 && code_k(c) == 4
+        @test distance(c, DistanceMIPAlgorithm(solver=HiGHS)) == 6
     end
 
     @testset "Weight-7 QLDPC codes" begin
@@ -150,5 +165,6 @@
         B = x + x^2 + y + z^2 + z^3
         c = two_block_group_algebra_codes(A, B)
         @test code_n(c) == 30 && code_k(c) == 4
+        @test distance(c, DistanceMIPAlgorithm(solver=HiGHS)) == 5
     end
 end