+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 
 ## v0.10.0 - dev
 
+- `QuantumClifffordJuMPExt`: New extension enabling:
+    - Compute minimum distance for quantum LDPC codes via Mixed Integer Programming (MIP) using `JuMP`.
+    - Introduce the `AbstractDistanceAlg` interface in `QECCore`.
 - Improvements to `GeneralizedStabilizer` API:
     - Tensor product operations:
         - Between multiple `GeneralizedStabilizer` states: `genstab₁ ⊗ genstab₂ ⊗ ...`
@@ -16,9 +19,10 @@
     - Product operations:
         - Between a `UnitaryPauliChannel` and a `GeneralizedStabilizer`: `pcT * genstab`
 - `QuantumClifford` now depends on `QECCore`. `QECCore` is a new package separate from `QuantumClifford`, which specifies the interfaces and several essential error correction codes. Here are some changes to the function and type names:
-  - `parity_check_x` -> `parity_matrix_x` and `parity_check_z` -> `parity_matrix_z`
+  - `parity_checks_x` -> `parity_matrix_x` and `parity_checks_z` -> `parity_matrix_z`
   - `AbstractECC` -> `AbstractQECC`
   - `ClassicalCode` -> `AbstractCECC`
+  - Some codes are moved to `QECCore` from `QuantumClifford`, including `Toric`, `Surface`, `RepCode`, `CSS`, `Shor9`, `Steane7`, `Cleve8`, `Perfect5`, `Bitflip3`.
 - **(breaking)** `StabMixture` was renamed to `GeneralizedStabilizer`.
 - **(fix)** `rowdecompose` was not accounting for the phase of the input Pauli string, leading to potential errors in nonclifford functionality.
 - `expect` is now implemented for `GeneralizedStabilizer`.